[up] Add thorough, generous error handling and admin mail notifications
Catch and log everything that goes wrong in any way.
For API (500ers), WEB (console.error, uncaughts, ...), SSR Server (ssr-plugin onError()) errors: Notify the admin via email and log the error
It would be great to use a proper logger and configure an automatic email transport for errors, so the mail server does not need to be called explicitly. But this can be done some other time
Add a small custom 500 html file for SSR errors, e.g. when API is dead
Move the noscript check back to index.html to avoid SSR hydration weirdness
Handle response codes in error handler, not in axios interceptor. In the latter, throw always if !== 200